refactor: moved Cards and Menu components to Shared directory
