+++ app/(components)/Navigation/Mobile/MobileNavigation.tsx
@@ -4,7 +4,7 @@ import { Collapse, Icon, Image, Link, useDisclosure } from "@chakra-ui/react"
 import * as React from "react"
 import { useContext } from "react"
 import { MobileNavbarContext } from "@/app/(components)/Navigation/Mobile/MobileNavbarProvider"
-import { ClickableCategoryProps, Item } from "@/app/(components)/Menu/ExpandableMenu"
+import { ClickableCategoryProps, Item } from "@/components/Shared/Menu/ExpandableMenu"
 import { ChevronDownIcon } from "@chakra-ui/icons"
