+++ app/(components)/Navigation/NavigationBar.tsx
@@ -1,7 +1,7 @@
 import ColorModeSwitcher from "@/app/(components)/ColorModeSwitcher"
 import * as React from "react"
 import SessionProfile from "@/app/(components)/Navigation/SessionProfile"
-import ClickableCategory, { ClickableCategoryProps, Item } from "@/app/(components)/Menu/ExpandableMenu"
+import ClickableCategory, { ClickableCategoryProps, Item } from "@/components/Shared/Menu/ExpandableMenu"
 import MobileNavbarProvider from "@/app/(components)/Navigation/Mobile/MobileNavbarProvider"
 import MobileNavigationBar from "@/app/(components)/Navigation/Mobile/MobileNavigation"
 import HamburgerMenu from "@/app/(components)/Navigation/Mobile/HamburgerMenu"
