+++ app/(components)/[study]/[semester]/LectureListItem.tsx
@@ -1,7 +1,7 @@
 import { BasicLecture } from "campus-scraper"
 import Link from "next/link"
-import { Text } from "@/app/(components)/ResponsiveTags/Text"
-import ServerImage from "@/app/(components)/ResponsiveTags/ServerImage"
+import { Text } from "@/components/Shared/Responsive/Text"
+import ServerImage from "@/components/Shared/Responsive/ServerImage"
 import User from "@/public/lectureIcons/user.svg"
