+++ app/(components)/[study]/[semester]/[id]/LectureDetails.tsx
@@ -1,8 +1,8 @@
 import { Lecture } from "campus-scraper"
 import { emptyLecture } from "@/app/api/data"
 import BackButton from "@/app/(components)/Navigation/BackButton"
-import Card from "@/app/(components)/Cards/Card"
-import { Text } from "@/app/(components)/ResponsiveTags/Text"
+import Card from "@/app/(components)/Shared/Cards/Card"
+import { Text } from "@/app/(components)/Shared/Responsive/Text"
 import Link from "next/link"
 import LectureDescription from "@/components/[study]/[semester]/[id]/LectureDescription"
 import Schedule from "@/components/[study]/[semester]/[id]/LectureSchedule"
